fix:修复一个蜜蜂的恶性bug 修复了WAP和转底炉不能制作的问题

+++ src/main/java/com/magicbee/ctnhmana/common/multiblock/ArcaneHighEnergyCompressionReactorCore.java
@@ -69,7 +69,7 @@ public class ArcaneHighEnergyCompressionReactorCore extends RecipeMultiblockMach
     protected final NotifiableItemStackHandler inventory;
     @Getter
     @Persisted
-    public long maxEU = Integer.MAX_VALUE;
+    public long maxEU = Long.MAX_VALUE;
     @Getter
     @Persisted
     public long EU = 0L;
